requirements.txt
discord.py[voice]
requests

# git+https://github.com/Rapptz/discord.py
# git+git://github.com/ExtraRandom/Game_Stores_API
# Pillow
# mcstatus
# beautifulsoup4